requirements.txt
Flask
openai-whisper
python-dotenv
requests